+++ R/cont_cdfplot.R
@@ -93,6 +93,7 @@
 #' @keywords survey plot
 #'
 #' @examples
+#' \dontrun{
 #' dframe <- data.frame(
 #'   siteID = paste0("Site", 1:100),
 #'   wgt = runif(100, 10, 100),
@@ -114,7 +115,6 @@
 #'   siteID = "siteID", weight = "wgt", xcoord = "xcoord", ycoord = "ycoord",
 #'   stratumID = "stratum", popsize = mypopsize
 #' )
-#' \dontrun{
 #' cont_cdfplot("myanalysis.pdf", myanalysis$CDF, ylab_r = "Stream Length (km)")
 #' }
 #'